back/requirements.txt
weaviate-client==4.11.1
fastapi
uvicorn[standard]
httpx>=0.26,<0.29
requests==2.31.0